Implement --force

+++ spec_parser/jsondump.py
@@ -2,6 +2,7 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
+import logging
 from pathlib import Path
 
 import jsonpickle
@@ -9,7 +10,11 @@ import jsonpickle
 
 def gen_jsondump(model, outdir, cfg):
     p = Path(outdir) / "jsondump"
-    p.mkdir()
+    if p.exists() and not cfg.opt_force:
+        logging.error(f"Destination for JSON dump: {p} already exists, will not overwrite")
+        return
+    p.mkdir(exist_ok=True)
 
     f = p / "model.json"
+
     f.write_text(jsonpickle.encode(model, indent=2, warn=True))
